+++ src/html_container.rs
@@ -5,8 +5,6 @@ use crate::content::BodyContent;
 use crate::Container;
 use crate::Html;
 use crate::Table;
-use std::collections::HashMap;
-use std::fmt::Display;
 
 /// An HTML element that can contain other HTML elements
 ///
@@ -15,8 +13,38 @@ use std::fmt::Display;
 pub trait HtmlContainer: Html + Sized {
     /// Adds the specified HTML element to this container
     ///
-    /// This method should *probably* not be used directly by client programs. For most
-    /// applications, the `add_x()` methods will provide more safety.
+    /// This method can be used as an escape hatch to insert arbitrary types into the HTML document,
+    /// helping to make up for those types which are not supported natively by this library. This
+    /// can be done by defining your own types that implement the [`Html`] trait.
+    ///
+    /// If you need a simple one-off, it may be more convenient to insert the element as a raw
+    /// string using [`add_raw`](HtmlContainer::add_raw) method instead
+    ///
+    /// # Example
+    /// ```
+    /// # use build_html::*;
+    /// #[derive(Debug)]
+    /// struct Span {
+    ///     content: String
+    /// }
+    ///
+    /// impl Span {
+    ///     pub fn new(content: impl ToString) -> Self {
+    ///         Span { content: content.to_string() }
+    ///     }
+    /// }
+    ///
+    /// impl Html for Span {
+    ///     fn to_html_string(&self) -> String {
+    ///         format!("<span>{}</span>", self.content)
+    ///     }
+    /// }
+    ///
+    /// let content = Container::default()
+    ///     .add_html(Box::new(Span::new("inner")))
+    ///     .to_html_string();
+    /// assert_eq!(content, "<div><span>inner</span></div>");
+    /// ```
     fn add_html(self, html: Box<dyn Html>) -> Self;
 
     /// Nest the specified container within this container
@@ -83,10 +111,10 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><h1>Header Text</h1></div>"#)
     /// ```
-    fn add_header<T: Display>(self, level: u8, text: T) -> Self {
+    fn add_header(self, level: u8, text: impl ToString) -> Self {
         let content = BodyContent::Header {
             level,
-            content: format!("{}", text),
+            content: text.to_string(),
             attr: Attributes::default(),
         };
         self.add_html(Box::new(content))
@@ -104,10 +132,14 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><h1 id="main-header">Header Text</h1></div>"#)
     /// ```
-    fn add_header_attr<T: Display>(self, level: u8, text: T, attr: HashMap<&str, &str>) -> Self {
+    fn add_header_attr<A, S>(self, level: u8, text: impl ToString, attr: A) -> Self
+    where
+        A: IntoIterator<Item = (S, S)>,
+        S: ToString,
+    {
         let content = BodyContent::Header {
             level,
-            content: format!("{}", text),
+            content: text.to_string(),
             attr: attr.into(),
         };
         self.add_html(Box::new(content))
@@ -124,10 +156,10 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><img src="myimage.png" alt="a test image"></div>"#)
     /// ```
-    fn add_image(self, src: &str, alt: &str) -> Self {
+    fn add_image(self, src: impl ToString, alt: impl ToString) -> Self {
         let content = BodyContent::Image {
-            src: src.into(),
-            alt: alt.into(),
+            src: src.to_string(),
+            alt: alt.to_string(),
             attr: Attributes::default(),
         };
         self.add_html(Box::new(content))
@@ -148,10 +180,14 @@ pub trait HtmlContainer: Html + Sized {
     ///     r#"<div><img src="myimage.png" alt="a test image" id="sample-image"></div>"#
     /// )
     /// ```
-    fn add_image_attr(self, src: &str, alt: &str, attr: HashMap<&str, &str>) -> Self {
+    fn add_image_attr<A, S>(self, src: impl ToString, alt: impl ToString, attr: A) -> Self
+    where
+        A: IntoIterator<Item = (S, S)>,
+        S: ToString,
+    {
         let content = BodyContent::Image {
-            src: src.into(),
-            alt: alt.into(),
+            src: src.to_string(),
+            alt: alt.to_string(),
             attr: attr.into(),
         };
         self.add_html(Box::new(content))
@@ -168,10 +204,10 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><a href="https://rust-lang.org/">Rust Homepage</a></div>"#)
     /// ```
-    fn add_link<T: Display>(self, href: &str, text: T) -> Self {
+    fn add_link(self, href: impl ToString, text: impl ToString) -> Self {
         let content = BodyContent::Link {
-            href: href.into(),
-            content: format!("{}", text),
+            href: href.to_string(),
+            content: text.to_string(),
             attr: Attributes::default(),
         };
         self.add_html(Box::new(content))
@@ -192,10 +228,14 @@ pub trait HtmlContainer: Html + Sized {
     ///     r#"<div><a href="https://rust-lang.org/" class="links">Rust Homepage</a></div>"#
     /// )
     /// ```
-    fn add_link_attr<T: Display>(self, href: &str, text: T, attr: HashMap<&str, &str>) -> Self {
+    fn add_link_attr<A, S>(self, href: impl ToString, text: impl ToString, attr: A) -> Self
+    where
+        A: IntoIterator<Item = (S, S)>,
+        S: ToString,
+    {
         let content = BodyContent::Link {
-            href: href.into(),
-            content: format!("{}", text),
+            href: href.to_string(),
+            content: text.to_string(),
             attr: attr.into(),
         };
         self.add_html(Box::new(content))
@@ -212,9 +252,9 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><p>This is sample paragraph text</p></div>"#)
     /// ```
-    fn add_paragraph<T: Display>(self, text: T) -> Self {
+    fn add_paragraph(self, text: impl ToString) -> Self {
         let content = BodyContent::Paragraph {
-            content: format!("{}", text),
+            content: text.to_string(),
             attr: Attributes::default(),
         };
         self.add_html(Box::new(content))
@@ -232,9 +272,13 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><p class="text">This is sample paragraph text</p></div>"#)
     /// ```
-    fn add_paragraph_attr<T: Display>(self, text: T, attr: HashMap<&str, &str>) -> Self {
+    fn add_paragraph_attr<A, S>(self, text: impl ToString, attr: A) -> Self
+    where
+        A: IntoIterator<Item = (S, S)>,
+        S: ToString,
+    {
         let content = BodyContent::Paragraph {
-            content: format!("{}", text),
+            content: text.to_string(),
             attr: attr.into(),
         };
         self.add_html(Box::new(content))
@@ -251,9 +295,9 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><pre>This | is   preformatted => text</pre></div>"#)
     /// ```
-    fn add_preformatted<T: Display>(self, text: T) -> Self {
+    fn add_preformatted(self, text: impl ToString) -> Self {
         let content = BodyContent::Preformatted {
-            content: format!("{}", text),
+            content: text.to_string(),
             attr: Attributes::default(),
         };
         self.add_html(Box::new(content))
@@ -271,11 +315,40 @@ pub trait HtmlContainer: Html + Sized {
     ///
     /// assert_eq!(content, r#"<div><pre id="code">This | is   preformatted => text</pre></div>"#)
     /// ```
-    fn add_preformatted_attr<T: Display>(self, text: T, attr: HashMap<&str, &str>) -> Self {
+    fn add_preformatted_attr<A, S>(self, text: impl ToString, attr: A) -> Self
+    where
+        A: IntoIterator<Item = (S, S)>,
+        S: ToString,
+    {
         let content = BodyContent::Preformatted {
-            content: format!("{}", text),
+            content: text.to_string(),
             attr: attr.into(),
         };
         self.add_html(Box::new(content))
+    }
+
+    /// Add raw content to this container. The content is pasted directly into the HTML
+    ///
+    /// This is intended to be used as an escape hatch for one-off insertions. If you want a more
+    /// reusable escape hatch, consider writing your own type implementing the [`Html`] trait. You
+    /// can then use [`add_html`](HtmlContainer::add_html) to insert boxed instances into the
+    /// container. See the documentation for that method for examples.
+    ///
+    /// # Example
+    /// ```
+    /// # use build_html::*;
+    /// let content = Container::default()
+    ///     .add_raw(r#"<video width="250"><source src="video.mp4" type="video/mp4"></video>"#)
+    ///     .to_html_string();
+    ///
+    /// assert_eq!(
+    ///     content,
+    ///     r#"<div><video width="250"><source src="video.mp4" type="video/mp4"></video></div>"#
+    /// );
+    /// ```
+    fn add_raw(self, content: impl ToString) -> Self {
+        self.add_html(Box::new(BodyContent::Raw {
+            content: content.to_string(),
+        }))
     }
 }