+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Changelog
 =========
 
+## Version 1.1.0 (TBD)
+* Rise of the Generics! -- `add_x` methods are now much more flexible!
+* Further defined escape hatches (for extensibility)
+* Added `add_raw` method for containers
+
 ## Version 1.0.0 (2021-01-24)
 * Initial Release